Extracts CalendarSwitch
